more styling cleanup
