+++ client/src/App.tsx
@@ -1,5 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import AppAppBar from './components/Home-Page/components/AppAppBar';
+import AppAppBar from './components/Home-Page/components/NavBar';
+import { CssBaseline, ThemeProvider } from '@mui/material';
 
 // import GoogleLoginButton from './components/Sign-In/components/GoogleLoginButton';
 // import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
