+++ client/src/components/Home-Page/Home.tsx
@@ -1,9 +1,9 @@
 import CssBaseline from '@mui/material/CssBaseline';
 import Divider from '@mui/material/Divider';
 import AppTheme from '../../shared-theme/AppTheme';
-import AppAppBar from './components/AppAppBar';
+import AppAppBar from './components/NavBar';
 import Hero from './components/Hero';
-import LogoCollection from './components/LogoCollection';
+// import LogoCollection from './components/LogoCollection';
 import FAQ from './components/FAQ';
 import Footer from './components/Footer';
 
@@ -15,7 +15,7 @@ export default function MarketingPage(props: { disableCustomTheme?: boolean }) {
       <AppAppBar />
       <Hero />
       <div>
-        <LogoCollection />
+        {/* <LogoCollection /> */}
         <Divider />
         <FAQ />
         <Divider />
