+++ client/src/components/Home-Page/components/Hero.tsx
@@ -3,6 +3,7 @@ import Button from '@mui/material/Button';
 import Container from '@mui/material/Container';
 import InputLabel from '@mui/material/InputLabel';
 import Link from '@mui/material/Link';
+import { Link as RouterLink } from 'react-router-dom';
 import Stack from '@mui/material/Stack';
 import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
@@ -125,10 +126,12 @@ export default function Hero() {
               }}
             />
             <Button
+              component={RouterLink}
+              to="/sign-up"
               variant="contained"
               color="primary"
               size="small"
-              sx={{ minWidth: 'fit-content' }}
+              sx={{ minWidth: 'fit-content'}}
             >
               Start now
             </Button>
